Termux: Work around shm symbol collision
Termux defines all of the shm functions as defines to libandroid-shmem
equivalents.
This messes with our function definitions that were using the same names
but weren't expecting them to be defines.

Changes our class function names to have an uppercase for the first
letter.

Changes our syscall enums to have an underscore at the start to not get
picked up by the preprocessor.